+++ micro_manager/adaptivity/adaptivity.py
@@ -6,6 +6,7 @@ from math import exp
 from typing import Callable
 from warnings import warn
 import subprocess
+import importlib
 from micro_manager.tools.logging_wrapper import Logger
 
 import numpy as np
@@ -30,8 +31,13 @@ class AdaptivityCalculator:
         self._hist_param = configurator.get_adaptivity_hist_param()
         self._adaptivity_data_names = configurator.get_data_for_adaptivity()
         self._adaptivity_type = configurator.get_adaptivity_type()
-        self._micro_file_name = configurator.get_micro_file_name()
-        self._lazy_init = configurator.initialize_sims_lazily()
+
+        self._micro_problem = getattr(
+            importlib.import_module(
+                configurator.get_micro_file_name(), "MicroSimulation"
+            ),
+            "MicroSimulation",
+        )
 
         self._coarse_tol = 0.0
         self._ref_tol = 0.0
@@ -40,8 +46,7 @@ class AdaptivityCalculator:
 
         # similarity_dists: 2D array having similarity distances between each micro simulation pair
         # This matrix is modified in place via the function update_similarity_dists
-        # NOTE: Data type restricted to float32 to save memory. Remove this restriction if higher precision is needed.
-        self._similarity_dists = np.zeros((nsims, nsims), dtype=np.float32)
+        self._similarity_dists = np.zeros((nsims, nsims))
 
         self._max_similarity_dist = 0.0
 
@@ -143,7 +148,8 @@ class AdaptivityCalculator:
         inactive_ids = np.where(self._is_sim_active == False)[0]
 
         # Start with a large distance to trigger the search for the most similar active sim
-        dist_min_start_value = 2 * self._max_similarity_dist
+        # Add the +1 for the case when the similarity distance matrix is zeros
+        dist_min_start_value = self._max_similarity_dist + 1
 
         # Associate inactive micro sims to active micro sims
         for inactive_id in inactive_ids: