+++ micro_manager/micro_manager.py
@@ -134,6 +134,7 @@ class MicroManagerCoupling(MicroManager):
             if self._is_adaptivity_with_load_balancing:
                 self._load_balancing_n = self._config.get_load_balancing_n()
 
+        self._adaptivity_n = self._config.get_adaptivity_n()
         self._adaptivity_output_n = self._config.get_adaptivity_output_n()
 
         # Define the preCICE Participant
@@ -166,38 +167,8 @@ class MicroManagerCoupling(MicroManager):
 
         dt = min(self._participant.get_max_time_step_size(), self._micro_dt)
 
-        if self._is_adaptivity_on:
-            # If micro simulations have been initialized, compute adaptivity before starting the coupling
-            if self._micro_sims_init or self._lazy_init:
-                self._logger.log_info_rank_zero(
-                    "Micro simulations have been initialized, so adaptivity will be computed before the coupling begins."
-                )
-
-                self._adaptivity_controller.compute_adaptivity(
-                    dt,
-                    self._micro_sims,
-                    self._data_for_adaptivity,
-                )
-            if self._lazy_init:
-                active_sim_ids = self._adaptivity_controller.get_active_sim_local_ids()
-                micro_problem = getattr(
-                    importlib.import_module(
-                        self._config.get_micro_file_name(), "MicroSimulation"
-                    ),
-                    "MicroSimulation",
-                )
-                for i in active_sim_ids:
-                    self._micro_sims[i] = create_simulation_class(micro_problem)(
-                        self._global_ids_of_local_sims[i]
-                    )
-                self._logger.log_info_rank_zero(
-                    "Some micro simulations have been initialized lazily before the start of the coupling."
-                )
-
         first_iteration = True
-        first_time_window = True
-
-        sim_states_cp = [None] * self._local_number_of_sims
+        checkpoints_saved = False
 
         while self._participant.is_coupling_ongoing():
 
@@ -209,9 +180,12 @@ class MicroManagerCoupling(MicroManager):
                         self._micro_sims[i].get_state() if self._micro_sims[i] else None
                     )
                 first_iteration = True
+                checkpoints_saved = True
 
             if self._is_adaptivity_on:
-                if self._adaptivity_in_every_implicit_step or first_iteration:
+                if (self._adaptivity_in_every_implicit_step or first_iteration) and (
+                    n % self._adaptivity_n == 0
+                ):
                     self._adaptivity_controller.compute_adaptivity(
                         dt,
                         self._micro_sims,
@@ -239,29 +213,12 @@ class MicroManagerCoupling(MicroManager):
 
                         self._has_sim_crashed = [False] * self._local_number_of_sims
 
-                # number_of_sims = self._comm.allgather(self._local_number_of_sims)
-                # assert self._global_number_of_sims == sum(number_of_sims)
-
-            # Write a checkpoint
-            if self._participant.requires_writing_checkpoint():
-                sim_states_cp.clear()
-                for i in range(self._local_number_of_sims):
-                    sim_states_cp.append(self._micro_sims[i].get_state())
-
-                if self._is_adaptivity_on:
-                    self._adaptivity_controller.write_checkpoint()
-
-                    active_sim_ids = (
-                        self._adaptivity_controller.get_active_sim_local_ids()
-                    )
+                    active_sim_ids = self._adaptivity_controller.get_active_sim_ids()
 
                     for active_id in active_sim_ids:
                         self._micro_sims_active_steps[active_id] += 1
 
-                        if (
-                            sim_states_cp[active_id] == None
-                            and self._participant.requires_writing_checkpoint()
-                        ):
+                        if sim_states_cp[active_id] == None and checkpoints_saved:
                             sim_states_cp[active_id] = self._micro_sims[
                                 active_id
                             ].get_state()
@@ -313,19 +270,18 @@ class MicroManagerCoupling(MicroManager):
                         self._micro_sims[i].set_state(sim_states_cp[i])
                 first_iteration = False
 
-                # If adaptivity is computed only once per time window, the states of sims need to be reset too
-                if self._is_adaptivity_on:
-                    if not self._adaptivity_in_every_implicit_step:
-                        self._adaptivity_controller.read_checkpoint()
-
-            if self._participant.is_time_window_complete():
+            if (
+                self._participant.is_time_window_complete()
+            ):  # Time window has converged, now micro output can be generated
                 if self._micro_sims_have_output:
                     if n % self._micro_n_out == 0:
                         for sim in self._micro_sims:
                             if sim:
                                 sim.output()
 
-                if self._is_adaptivity_on and n % self._adaptivity_output_n == 0:
+                if self._is_adaptivity_on and (
+                    n % self._adaptivity_output_n == 0 or n == 1
+                ):
                     self._adaptivity_controller.log_metrics(n)
 
                 if self._output_memory_usage:
@@ -333,14 +289,18 @@ class MicroManagerCoupling(MicroManager):
 
                 self._logger.log_info_rank_zero("Time window {} converged.".format(n))
 
-        mem_usage_output_file = (
-            self._output_dir + "mem_usage_" + str(self._rank) + ".csv"
-        )
-        with open(mem_usage_output_file, mode="w", newline="") as file:
-            writer = csv.writer(file)
-            writer.writerow(["Time window", "RSS (MB)"])
-            for i, rss_mb in enumerate(mem_usage):
-                writer.writerow([i, rss_mb])
+        if self._is_adaptivity_on and n % self._adaptivity_output_n != 0:
+            self._adaptivity_controller.log_metrics(n)
+
+        if self._output_memory_usage:
+            mem_usage_output_file = (
+                self._output_dir + "mem_usage_" + str(self._rank) + ".csv"
+            )
+            with open(mem_usage_output_file, mode="w", newline="") as file:
+                writer = csv.writer(file)
+                writer.writerow(["Time window", "RSS (MB)"])
+                for i, rss_mb in enumerate(mem_usage):
+                    writer.writerow([i, rss_mb])
 
         self._participant.finalize()
 
@@ -528,63 +488,86 @@ class MicroManagerCoupling(MicroManager):
         # Read initial data from preCICE, if it is available
         initial_data = self._read_data_from_precice(dt=0)
 
+        first_id = 0  # 0 if lazy initialization is off, otherwise the first active simulation ID
+        micro_sims_to_init = range(
+            1, self._local_number_of_sims
+        )  # All sims if lazy init is off, otherwise all active simulations
+
         if not initial_data:
             is_initial_data_available = False
-            if self._lazy_init:
-                raise Exception(
-                    "No initial macro data available, lazy initialization would result in only one active simulation."
-                )
         else:
             is_initial_data_available = True
+            if (
+                self._lazy_init
+            ):  # For lazy initialization, compute adaptivity with the initial macro data
+                for i in range(self._local_number_of_sims):
+                    for name in self._adaptivity_macro_data_names:
+                        self._data_for_adaptivity[name][i] = initial_data[i][name]
+
+                self._adaptivity_controller.compute_adaptivity(
+                    self._micro_dt, self._micro_sims, self._data_for_adaptivity
+                )
+
+                active_sim_ids = self._adaptivity_controller.get_active_sim_ids()
+
+                if active_sim_ids.size == 0:
+                    self._logger.log_info(
+                        "There are no active simulations on this rank."
+                    )
+                    return
+
+                for i in active_sim_ids:
+                    self._micro_sims[i] = create_simulation_class(micro_problem)(
+                        self._global_ids_of_local_sims[i]
+                    )
+
+                first_id = active_sim_ids[0]  # First active simulation ID
+                micro_sims_to_init = (
+                    active_sim_ids  # Only active simulations will be initialized
+                )
 
         # Boolean which states if the initialize() method of the micro simulation requires initial data
-        is_initial_data_required = False
+        sim_requires_init_data = False
 
         # Check if provided micro simulation has an initialize() method
         if hasattr(micro_problem, "initialize") and callable(
             getattr(micro_problem, "initialize")
         ):
-            if self._lazy_init:
-                self._logger.log_warning(
-                    "The initialize function of micro simulations will not be called when using "
-                    "lazy initialization and adaptivity can't use data returned by it."
-                )
-            else:
-                self._micro_sims_init = True  # Starting value before setting
+            self._micro_sims_init = True  # Starting value before setting
 
-                try:  # Try to get the signature of the initialize() method, if it is written in Python
-                    argspec = inspect.getfullargspec(micro_problem.initialize)
-                    if (
-                        len(argspec.args) == 1
-                    ):  # The first argument in the signature is self
-                        is_initial_data_required = False
-                    elif len(argspec.args) == 2:
-                        is_initial_data_required = True
-                    else:
+            try:  # Try to get the signature of the initialize() method, if it is written in Python
+                argspec = inspect.getfullargspec(micro_problem.initialize)
+                if (
+                    len(argspec.args) == 1
+                ):  # The first argument in the signature is self
+                    sim_requires_init_data = False
+                elif len(argspec.args) == 2:
+                    sim_requires_init_data = True
+                else:
+                    raise Exception(
+                        "The initialize() method of the Micro simulation has an incorrect number of arguments."
+                    )
+            except TypeError:
+                self._logger.log_info_rank_zero(
+                    "The signature of initialize() method of the micro simulation cannot be determined. Trying to determine the signature by calling the method."
+                )
+                # Try to get the signature of the initialize() method, if it is not written in Python
+                try:  # Try to call the initialize() method without initial data
+                    self._micro_sims[first_id].initialize()
+                    sim_requires_init_data = False
+                except TypeError:
+                    self._logger.log_info_rank_zero(
+                        "The initialize() method of the micro simulation has arguments. Attempting to call it again with initial data."
+                    )
+                    try:  # Try to call the initialize() method with initial data
+                        self._micro_sims[first_id].initialize(initial_data[first_id])
+                        sim_requires_init_data = True
+                    except TypeError:
                         raise Exception(
                             "The initialize() method of the Micro simulation has an incorrect number of arguments."
                         )
-                except TypeError:
-                    self._logger.log_info_rank_zero(
-                        "The signature of initialize() method of the micro simulation cannot be determined. Trying to determine the signature by calling the method."
-                    )
-                    # Try to get the signature of the initialize() method, if it is not written in Python
-                    try:  # Try to call the initialize() method without initial data
-                        self._micro_sims[0].initialize()
-                        is_initial_data_required = False
-                    except TypeError:
-                        self._logger.log_info_rank_zero(
-                            "The initialize() method of the micro simulation has arguments. Attempting to call it again with initial data."
-                        )
-                        try:  # Try to call the initialize() method with initial data
-                            self._micro_sims[0].initialize(initial_data[0])
-                            is_initial_data_required = True
-                        except TypeError:
-                            raise Exception(
-                                "The initialize() method of the Micro simulation has an incorrect number of arguments."
-                            )
 
-        if is_initial_data_required and not is_initial_data_available:
+        if sim_requires_init_data and not is_initial_data_available:
             raise Exception(
                 "The initialize() method of the Micro simulation requires initial data, but no initial macro data has been provided."
             )
@@ -593,10 +576,12 @@ class MicroManagerCoupling(MicroManager):
         if self._micro_sims_init:
 
             # Call initialize() method of the micro simulation to check if it returns any initial data
-            if is_initial_data_required:
-                initial_micro_output = self._micro_sims[0].initialize(initial_data[0])
+            if sim_requires_init_data:
+                initial_micro_output = self._micro_sims[first_id].initialize(
+                    initial_data[first_id]
+                )
             else:
-                initial_micro_output = self._micro_sims[0].initialize()
+                initial_micro_output = self._micro_sims[first_id].initialize()
 
             if (
                 initial_micro_output is None
@@ -607,28 +592,35 @@ class MicroManagerCoupling(MicroManager):
                 )
                 self._micro_sims_init = False
 
-                if is_initial_data_required:
-                    for i in range(1, self._local_number_of_sims):
+                if sim_requires_init_data:
+                    for i in micro_sims_to_init:
                         self._micro_sims[i].initialize(initial_data[i])
                 else:
-                    for i in range(1, self._local_number_of_sims):
+                    for i in micro_sims_to_init:
                         self._micro_sims[i].initialize()
             else:  # Case where the initialize() method returns data
                 if self._is_adaptivity_on:
+                    initial_micro_data: dict[str, list] = dict()
+
+                    for name in initial_micro_output.keys():
+                        initial_micro_data[name] = [0] * self._local_number_of_sims
+                        # Save initial data from first micro simulation as we anyway have it
+                        initial_micro_data[name][first_id] = initial_micro_output[name]
+
                     # Save initial data from first micro simulation as we anyway have it
                     for name in initial_micro_output.keys():
                         if name in self._data_for_adaptivity:
-                            self._data_for_adaptivity[name][0] = initial_micro_output[
-                                name
-                            ]
+                            self._data_for_adaptivity[name][
+                                first_id
+                            ] = initial_micro_output[name]
                         else:
                             raise Exception(
                                 "The initialize() method needs to return data which is required for the adaptivity calculation."
                             )
 
                     # Gather initial data from the rest of the micro simulations
-                    if is_initial_data_required:
-                        for i in range(1, self._local_number_of_sims):
+                    if sim_requires_init_data:
+                        for i in micro_sims_to_init:
                             initial_micro_output = self._micro_sims[i].initialize(
                                 initial_data[i]
                             )
@@ -636,18 +628,34 @@ class MicroManagerCoupling(MicroManager):
                                 self._data_for_adaptivity[name][
                                     i
                                 ] = initial_micro_output[name]
+                                initial_micro_data[name][i] = initial_micro_output[name]
                     else:
-                        for i in range(1, self._local_number_of_sims):
+                        for i in micro_sims_to_init:
                             initial_micro_output = self._micro_sims[i].initialize()
                             for name in self._adaptivity_micro_data_names:
                                 self._data_for_adaptivity[name][
                                     i
                                 ] = initial_micro_output[name]
+                                initial_micro_data[name][i] = initial_micro_output[name]
+
+                    if (
+                        self._lazy_init
+                    ):  # If lazy initialization is on, initial states of inactive simulations need to be determined
+                        self._adaptivity_controller.get_full_field_micro_output(
+                            initial_micro_data
+                        )
+                        for i in range(self._local_number_of_sims):
+                            for name in self._adaptivity_micro_data_names:
+                                self._data_for_adaptivity[name][i] = initial_micro_data[
+                                    name
+                                ][i]
+                        del initial_micro_data  # Once the initial data is fed into the adaptivity data, it is no longer required
+
                 else:
                     self._logger.log_warning_rank_zero(
                         "The initialize() method of the Micro simulation returns initial data, but adaptivity is turned off. The returned data will be ignored. The initialize method will nevertheless still be called."
                     )
-                    if is_initial_data_required:
+                    if sim_requires_init_data:
                         for i in range(1, self._local_number_of_sims):
                             self._micro_sims[i].initialize(initial_data[i])
                     else: